# BFGS: Performance Improvements