--- a/src/primitives/circle.rs
+++ b/src/primitives/circle.rs
@@ -70,8 +70,8 @@ impl PrimitiveLike for Circle {
         self.gradient = SVector::<f64, 1>::zeros();
     }
 
-    fn get_data(&self) -> DVector<f64> {
-        DVector::from_row_slice(self.data.as_slice())
+    fn get_data(&self) -> DVectorView<f64> {
+        self.data.as_view()
     }
 
     fn set_data(&mut self, data: DVectorView<f64>) {
@@ -79,8 +79,8 @@ impl PrimitiveLike for Circle {
         self.data.copy_from(&data);
     }
 
-    fn get_gradient(&self) -> DVector<f64> {
-        DVector::from_row_slice(self.gradient.as_slice())
+    fn get_gradient(&self) -> DVectorView<f64> {
+        self.gradient.as_view()
     }
 
     fn to_primitive(&self) -> super::Primitive {