# Overview

Improves performance by reducing allocations in gradient gathering by returning views into the data/gradient in each primitive. BFGS performance is significantly improved by avoiding the formation of large matrix products.

# Details

The original version of the BFGS update repeatedly formed intermediate products for terms such as `h * y`. That product is computed once and reused. Additionally, updates of the form `alpha * x * y^T + beta * A` (for vectors `x`, `y`, matrix `A`, and scalars `alpha`, `beta`) can be computed without forming the intermediate product `x * y^T`. This is implemented using the [ger](https://docs.rs/nalgebra/latest/nalgebra/base/struct.Matrix.html#method.ger) function in `nalgebra`.

# Benchmarks

On my machine:
```
Benchmark: CirclesWithLines
n:   30,        primitives:   59,       constraints:  88,       solver: BFGSSolver         ,    solved: true,   error: 0.00,    duration: 0ms
n:   50,        primitives:   99,       constraints: 148,       solver: BFGSSolver         ,    solved: true,   error: 0.00,    duration: 1ms
n:  100,        primitives:  199,       constraints: 298,       solver: BFGSSolver         ,    solved: true,   error: 0.00,    duration: 6ms
n:  300,        primitives:  599,       constraints: 898,       solver: BFGSSolver         ,    solved: true,   error: 0.00,    duration: 48ms
Benchmark: StairsWithLines
n:   30,        primitives:   59,       constraints:  59,       solver: BFGSSolver         ,    solved: true,   error: 0.00,    duration: 0ms
n:   50,        primitives:   99,       constraints:  99,       solver: BFGSSolver         ,    solved: true,   error: 0.00,    duration: 1ms
n:  100,        primitives:  199,       constraints: 199,       solver: BFGSSolver         ,    solved: true,   error: 0.00,    duration: 10ms
n:  300,        primitives:  599,       constraints: 599,       solver: BFGSSolver         ,    solved: true,   error: 0.00,    duration: 245ms
```