@jplatte 2024-05-30
I see you posted some benchmark results, but I don't see any benchmarks in this repo or in your PR.

Maybe this is not helpful, but just wanted to drop the info that Cargo has (limited) built-in support for running benchmarks. You still need to pull in an external library to actually define the benchmarks, but there's two pretty good crates to choose from: [criterion](https://bheisler.github.io/criterion.rs/book/criterion_rs.html), [divan](https://docs.rs/divan/latest/divan/).