@dwmunster 2024-05-30
> but I don't see any benchmarks in this repo or in your PR.

The benchmarks have currently been implemented as a set of [(ignored) tests](/src/examples/benchmarks/mod.rs). I've done some tire kicking of criterion locally but haven't had a chance to look at pulling that in and/or integration with actions, though [I agree](https://github.com/CADmium-Co/ISOtope/pull/19#issuecomment-2138327873) that would be valuable.